Adding a new webpage with links and images
Adds a new webpage that has a picture of a dog on the front page, and then a link to the about page of the odin project, and a link to the about page for the website. The about page also has the picture of the dog, using ../ for the relative link to the image.